+++ pinto/attributes/default.rb
@@ -1,15 +1,13 @@
-default.pinto.bootstrap.user = 'root'
-default.pinto.bootstrap.group = 'root'
+default.pinto.bootstrap.user = 'pinto'
+default.pinto.bootstrap.group = 'pinto'
 
-default.pinto.bootstrap.repo_url = 'http://stratopan.com:2700/Stratopan/Pinto/Production'
 default.pinto.bootstrap.installer_url = 'http://getpinto.stratopan.com'
 
-default.pinto.bootstrap.home = "#{ENV['HOME']}/opt/local/pinto"
-
+default.pinto.bootstrap.home = "/home/pinto/opt/local/pinto"
 
 default.pinto.bootstrap.cpanminus_url = 'http://xrl.us/cpanm'
 
-default.pinto.bootstrap.packages = %w[ ]
+default.pinto.bootstrap.packages = %w[ curl ]
 default.pinto.bootstrap.missed.cpan.packages = %w[ Time::HiRes CGI Module::CoreList ]
 
 case platform 
@@ -24,7 +22,7 @@ case platform
 end
 
 
-default.pinto.server.repo_root = "#{ENV['HOME']}/opt/local/pinto/repo/"
+default.pinto.server.repo_root = "/home/pinto/opt/local/pinto/repo/"
 default.pinto.server.host = '0.0.0.0'
 default.pinto.server.port = '5000'
 default.pinto.server.workers = '3'